Add a summary view when nodes are collapsed.  Fix a DB bug with deleting legos.  Fix a deadlock bug with saving legos.  Add support for text and boolean in the tree view under value and destination.  Clean up the styling of the concept view node.  Changing some other styling.

+++ src/main/java/gov/va/legoEdit/model/bdbModel/LegoListBDB.java
@@ -71,7 +71,7 @@ public class LegoListBDB
             Integer temp = legoUUIDsUsage.get(l.getLegoUUID());
             if (temp == null)
             {
-                temp = new Integer(0);
+                temp = new Integer(1);
             }
             else
             {
@@ -94,7 +94,7 @@ public class LegoListBDB
         Integer temp = legoUUIDsUsage.get(lego.getLegoUUID());
         if (temp == null)
         {
-            temp = new Integer(0);
+            temp = new Integer(1);
         }
         else
         {
